finetune-model/requirements.txt
# Enhanced Fine-tuning Framework Requirements
# Core ML libraries
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers==4.41.2
peft==0.10.0
accelerate==0.30.1
datasets==2.19.1
trl==0.8.6
bitsandbytes>=0.41.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0

# Optional: GPU monitoring (uncomment if needed)
# nvidia-ml-py>=11.0.0
# psutil>=5.9.0

# Development and testing
requests>=2.28.0

